% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDE_helper_functions.R
\name{classifyDE}
\alias{classifyDE}
\title{Classify DE result}
\usage{
classifyDE(de, param, ctr, overall_ctr)
}
\arguments{
\item{de}{An object of ("trimmed") results in the form as returned by scrapeDEresults.}

\item{param}{A data.frame of parameters used during classification.}

\item{ctr}{A data.frame of contrasts used for constructing DE.}

\item{overall_ctr}{A data.frame of overall contrasts}
}
\value{
A data.frame with classified results.
}
\description{
Given DE data for various contrasts (\code{ctr}), each of the contrasts will be classified as lethal, beneficial
or neutral from l2fc and p-value based on parameters passed in from \code{param} (\code{l_thr}, \code{p_thr}).
Based on a group of (overall) contrasts (\code{overall_ctr}) number of beneficial, lethal or neutral shRNAs in a group of
contrasts is counted and classified as lethal, beneficial and neutral using \code{overall_thr} parameter.
}
