% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDE_helper_functions.R
\name{countClassified}
\alias{countClassified}
\title{Count classified results}
\usage{
countClassified(count, ctr, overall_ctr, classy)
}
\arguments{
\item{count}{Data.frame}

\item{ctr}{Data.frame of contrasts for which counting should be done.}

\item{overall_ctr}{Data.frame of overall contrasts.}

\item{classy}{Which classification to count. Possible options are \code{lethal} and \code{beneficial}.}
}
\value{
Matrix of counts where rows are genes and columns are contrasts.
}
\description{
For each contrast, count how many times neutral, lethal or beneficial species comes up.
}
