% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDE_helper_functions.R
\name{differentialExpression}
\alias{differentialExpression}
\title{Perform differential expression based on specified design}
\usage{
differentialExpression(contrasts, sample_key, data, design = ~0 + group +
  replicate)
}
\arguments{
\item{contrasts}{A vector of contrasts (length of 2) for group 1 and group 2.}

\item{sample_key}{Data.frame. A data.frame as returned by \code{importSampleKey}.}

\item{data}{Count data from \code{importCountData}.}

\item{design}{Design used to perform differential expression.}
}
\value{
A list of DESeq results as returned by \code{DESeq2::results}.
}
\description{
Based on provided contrasts, provide differential expression.
}
