% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDE.R
\name{doDE}
\alias{doDE}
\title{Make differential expression analysis}
\usage{
doDE(input, sample_list = NULL)
}
\arguments{
\item{input}{Character. A relative or absolute path to a parameters file. See Details.}

\item{sample_list}{Character. Vector of absolute or relative paths to sample expressions or
folder containing sample expressions. See \code{importExpressionData} for more info.}
}
\value{
Files are returned: \code{deseq_results.txt}, \code{class_results.txt},
\code{beneficial_counts.txt}, \code{lethal_counts.txt}.
For usage, check out test file.
}
\description{
Input a set of count matrices, define relationships and perform a differential analysis.
Return several outputs, including differentil expression result and some extra  comparisons.
}
\details{
\code{input} file should be an excel (xlsx) file with tabs sample_key, contrasts,
overall_contrasts and classification_parameters. \code{sample_key} should have columns
sample, <variables> (depends on the experiment) and replicate. \code{contrasts} should
have columns group_1 and group_2, and they should hold which groups to test against each
other. Sheet \code{overall_contrasts} should have groups of overall contrasts in columns,
where rows represent pairs of comparisons. \code{classification_parameters} sheet should
have twothree columns, threshold, value and description. See sample file for details.
}
