% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDE_helper_functions.R
\name{fetchGeneOrHairpin}
\alias{fetchGeneOrHairpin}
\title{Process string to extract shRNA or gene}
\usage{
fetchGeneOrHairpin(x, entity = c("shrna", "gene"))
}
\arguments{
\item{x}{Character. A vector of shRNA-gene names.}

\item{entity}{Character. Return either shRNA ("shrna") or gene ("gene").}
}
\value{
Depending on which \code{entity} is chosen, shRNA part or gene part are returned.
}
\description{
shRNA-gene character is split on "-" and gene or shRNA are returned based on parameter entity.
}
