% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDE_helper_functions.R
\name{fetchTHvalue}
\alias{fetchTHvalue}
\title{Return specified threshold value.}
\usage{
fetchTHvalue(x, th = NULL)
}
\arguments{
\item{x}{Data.frame. A threshold data.frame as returned by importClassificationParams.}

\item{th}{Character. Character of length 1. Possible values are \code{l_thr}, \code{p_thr},
\code{overall_thr}.}
}
\value{
A value corresponding to \code{th}.
}
\description{
A helper function which makes fetching of parameters easy, pretty and extensible.
}
