% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDE_helper_functions.R
\name{importClassificationParams}
\alias{importClassificationParams}
\title{Import parameters used for classification}
\usage{
importClassificationParams(x, trim = TRUE)
}
\arguments{
\item{x}{Character. Absolute or relative path to an .xlsx file. See \code{doDE} description
for more information.}

\item{trim}{Logical. If \code{TRUE} (default), description column will be omitted.}
}
\value{
A data.frame with two columns, one being name of the parameters and second being its value.
}
\description{
Import parameters used for classification
}
