% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDE_helper_functions.R
\name{importContrastList}
\alias{importContrastList}
\title{Import contrast list}
\usage{
importContrastList(x)
}
\arguments{
\item{x}{Character. Absolute or relative path to an .xlsx file. See \code{doDE} description
for more information.}
}
\value{
A data.frame of contrasts for group_1 and group_2 and a concatenation of
group_1.group_2 as column \code{contrast}.
}
\description{
User input should be contrasts (group_1, group_2) which will be used to compare
in the DE run.
}
