% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDE_helper_functions.R
\name{importExpressionsData}
\alias{importExpressionsData}
\title{Import expression data based on sample key.}
\usage{
importExpressionsData(sample_key, sample_list = NULL)
}
\arguments{
\item{sample_key}{A data.frame with columns sample_name, <groups>* and replicate. Number of
columns for groups may vary.}

\item{sample_list}{Character. Vector of absolute or relative paths to sample count matrices. If
NULL, subfolder ./data is search for all files ending in .txt.}
}
\value{
A data.frame of shRNA counts for selected samples.
}
\description{
Imports expression data based on sample name specified in \code{sample_key}. It takes sample
name from \code{sample_key} and tries to match the name in \code{sample_list}. For final
output, results are merged based on shRNA.
}
