% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDE_helper_functions.R
\name{importOverallContrasts}
\alias{importOverallContrasts}
\title{Import overall contrasts}
\usage{
importOverallContrasts(x)
}
\arguments{
\item{x}{Character. Absolute or relative path to an .xlsx file. See \code{doDE} description
for more information.}
}
\value{
A data.frame of overall contrasts. Each column represents a group of contrasts.
}
\description{
From parameter file's \code{overall_contrasts} fetch all defined overall contrasts.
}
