% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDE_helper_functions.R
\name{importSampleKey}
\alias{importSampleKey}
\title{Import sample key}
\usage{
importSampleKey(x)
}
\arguments{
\item{x}{Character. Absolute or relative path to an .xlsx file. See \code{doDE} description
for more information.}
}
\value{
A data.frame with columns sample, replicate and group, used in differential expression.
}
\description{
Take in parameter file (.xlsx) and from tab \code{sample_key} extract information on study design.
}
