% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLibrary.R
\name{makeLibrary}
\alias{makeLibrary}
\title{Create library of shRNA}
\usage{
makeLibrary(input, output = NULL)
}
\arguments{
\item{input}{A raw file of three columns where first value is shRNA, second value is gene
name and third value is sequence.}

\item{output}{Character. Optional. If a character string of length one is supplied, it is
stripped of file extension and appended \code{fasta}. If \code{NULL}, name of input is used,
where \code{.fasta} extension is appended and written to working directory.}
}
\value{
Returns link to file of the library, which has been written to working directory or
path specified in \code{output} as a side effect.
}
\description{
Take a raw file and convert it to a valid fasta format. This file is necessary for performing
alignment.
}
\author{
Roman Luštrik (roman@genialis.com) based on original code by Nicholas Neill (nicholas.neill@bcm.edu).
}
