% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDE_helper_functions.R
\name{scrapeDEresults}
\alias{scrapeDEresults}
\title{Scrape DESeq2 results and format in reusable format}
\usage{
scrapeDEresults(dds, shrna, gene)
}
\arguments{
\item{dds}{Data.frame object from \code{differentialExpression}.}

\item{shrna}{Character. Character of shRNA names.}

\item{gene}{Character. Character of gene names.}
}
\value{
A data.frame of shRNA, gene names and DESeq2 results.
}
\description{
A handy function which will extract information from `differentialExpression` result into a more user
friendly form.
}
