% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDE_helper_functions.R
\name{trimResults}
\alias{trimResults}
\title{Trim results of NAs.}
\usage{
trimResults(de)
}
\arguments{
\item{de}{Data.frame. Object from \code{scrapeDEresults}.}
}
\value{
A possible subset of \code{de} where shRNA (rows) for all samples with missing expression have been
removed.
}
\description{
Will remove shRNas which have missing values on at least one contrast.
}
